The "conftest" and "test_create_order_authorized_user_success" are fixed.

+++ tests/conftest.py
@@ -14,15 +14,18 @@ def default_user():
     yield payload
 
     response = UserApi.login_user(payload)
-    token = response.json()['accessToken']
-    requests.delete(BASE_URL + DELETE_USER_ENDPOINT, headers={'authorization': token})
+    if response.status_code == 200:
+        token = response.json()['accessToken']
+        requests.delete(BASE_URL + DELETE_USER_ENDPOINT, headers={'authorization': token})
+    else:
+        pass
+
 
 # Фикстура - получение токена авторизации после регистрации и авторизации рандомного пользователя
 @pytest.fixture(scope='function')
-def default_update_user():
-    payload = REGISTER_USER_BODY
-    UserApi.register_user(payload)
-    login_body = helpers.get_login_user_data(payload)
+def default_update_user(default_user):
+    UserApi.register_user(default_user)
+    login_body = helpers.get_login_user_data(default_user)
     response_login_user = UserApi.login_user(login_body)
     token = response_login_user.json()['accessToken']
     yield token
